+++ src/messaging/index.js
@@ -4,21 +4,21 @@
 var async = require('async');
 var validator = require('validator');
 
-var db = require('./database');
-var user = require('./user');
-var plugins = require('./plugins');
-var meta = require('./meta');
-var utils = require('./utils');
+var db = require('../database');
+var user = require('../user');
+var plugins = require('../plugins');
+var meta = require('../meta');
+var utils = require('../utils');
 
 var Messaging = module.exports;
 
-require('./messaging/data')(Messaging);
-require('./messaging/create')(Messaging);
-require('./messaging/delete')(Messaging);
-require('./messaging/edit')(Messaging);
-require('./messaging/rooms')(Messaging);
-require('./messaging/unread')(Messaging);
-require('./messaging/notifications')(Messaging);
+require('./data')(Messaging);
+require('./create')(Messaging);
+require('./delete')(Messaging);
+require('./edit')(Messaging);
+require('./rooms')(Messaging);
+require('./unread')(Messaging);
+require('./notifications')(Messaging);
 
 
 Messaging.getMessages = function (params, callback) {
@@ -384,4 +384,4 @@ Messaging.hasPrivateChat = function (uid, withUid, callback) {
 	], callback);
 };
 
-Messaging.async = require('./promisify')(Messaging);
+Messaging.async = require('../promisify')(Messaging);
