+++ src/plugins/index.js
@@ -12,10 +12,10 @@ var middleware;
 
 var Plugins = module.exports;
 
-require('./plugins/install')(Plugins);
-require('./plugins/load')(Plugins);
-require('./plugins/hooks')(Plugins);
-Plugins.data = require('./plugins/data');
+require('./install')(Plugins);
+require('./load')(Plugins);
+require('./hooks')(Plugins);
+Plugins.data = require('./data');
 
 Plugins.getPluginPaths = Plugins.data.getPluginPaths;
 Plugins.loadPluginInfo = Plugins.data.loadPluginInfo;
@@ -113,7 +113,7 @@ Plugins.reload = function (callback) {
 };
 
 Plugins.reloadRoutes = function (router, callback) {
-	var controllers = require('./controllers');
+	var controllers = require('../controllers');
 	Plugins.fireHook('static:app.load', { app: app, router: router, middleware: middleware, controllers: controllers }, function (err) {
 		if (err) {
 			winston.error('[plugins] Encountered error while executing post-router plugins hooks', err);
@@ -243,7 +243,7 @@ Plugins.normalise = function (apiReturn, callback) {
 	});
 };
 
-Plugins.nodeModulesPath = path.join(__dirname, '../node_modules');
+Plugins.nodeModulesPath = path.join(__dirname, '../../node_modules');
 
 Plugins.showInstalled = function (callback) {
 	var pluginNamePattern = /^(@.*?\/)?nodebb-(theme|plugin|widget|rewards)-.*$/;
